backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
python-jose[cryptography]
passlib[argon2]
bcrypt==4.0.1
python-multipart
pydantic[email]
sendgrid
stripe
httpx
supabase
colorama
